+++ packages/shared/src/components/cards/squad/SquadList.tsx
@@ -1,9 +1,5 @@
-import React, {
-  AnchorHTMLAttributes,
-  HTMLAttributes,
-  ReactElement,
-  ReactNode,
-} from 'react';
+import React, { ReactElement } from 'react';
+import { useRouter } from 'next/router';
 import { Squad } from '../../../graphql/sources';
 import {
   Typography,
@@ -12,49 +8,31 @@ import {
 } from '../../typography/Typography';
 import { Separator } from '../common';
 import { largeNumberFormat } from '../../../lib';
-import { SquadJoinButtonWrapper } from './common/SquadJoinButton';
-import { ButtonVariant } from '../../buttons/common';
-import { SquadCardAction } from './common/types';
 import { ArrowIcon } from '../../icons';
 import { IconSize } from '../../Icon';
 import { CardLink } from '../Card';
-import { SquadImage } from './common/SquadImage';
+import { SquadJoinButton } from '../../squads/SquadJoinButton';
+import { Origin } from '../../../lib/log';
 
-interface SquadListBaseProps {
+interface SquadListProps {
   squad: Squad;
   isUserSquad?: boolean;
-  elementProps?:
-    | HTMLAttributes<HTMLSpanElement>
-    | AnchorHTMLAttributes<HTMLAnchorElement>;
-  icon?: ReactNode;
 }
-
-interface UserSquadProps extends SquadListBaseProps {
-  isUserSquad: true;
-  action?: never; // action is not allowed when isUserSquad is true
-}
-
-interface NonUserSquadProps extends SquadListBaseProps {
-  isUserSquad?: false;
-  action: SquadCardAction; // action is required when isUserSquad is false
-}
-
-type SquadListProps = UserSquadProps | NonUserSquadProps;
 
 export const SquadList = ({
   squad,
-  action,
   isUserSquad,
-  icon,
 }: SquadListProps): ReactElement => {
+  const router = useRouter();
+  const { image, name, permalink } = squad;
+
   return (
     <div className="relative flex flex-row items-center gap-4">
-      <CardLink href={squad.permalink} rel="noopener" title={squad.name} />
-      <SquadImage
-        image={squad?.image}
-        icon={icon}
-        title={squad.name}
-        size="size-14"
+      <CardLink href={permalink} rel="noopener" title={name} />
+      <img
+        className="size-14 rounded-full"
+        src={image}
+        alt={`${name} source`}
       />
       <div className="flex w-0 flex-grow flex-col">
         <Typography type={TypographyType.Callout} bold truncate>
@@ -81,13 +59,13 @@ export const SquadList = ({
           size={IconSize.Small}
         />
       ) : (
-        <SquadJoinButtonWrapper
-          action={action}
-          source={squad}
-          variant={ButtonVariant.Float}
-          className={{
-            squadJoinButton: '!btn-tertiaryFloat',
-          }}
+        <SquadJoinButton
+          className={{ button: '!btn-tertiaryFloat z-0' }}
+          squad={squad}
+          origin={Origin.SquadDirectory}
+          onSuccess={() => router.push(permalink)}
+          joinText="Join"
+          data-testid="squad-action"
         />
       )}
     </div>
